dev-requirements.txt
Flask-Testing==0.4.2
httmock==1.2.5
coveralls==0.5
selenium==3.5.0

